+++ src/store/tag.ts
@@ -6,6 +6,7 @@
 
 import { defineStore } from 'pinia'
 import { firstUpperCase } from '/@/transforms/text'
+import { UniversalExtend } from '/@/constants/state'
 import nodepress from '/@/services/nodepress'
 
 export type TagMap = Map<string, Tag>
@@ -18,7 +19,7 @@ export interface Tag {
   description: string
   update_at: string
   create_at: string
-  extends: $TODO[]
+  extends?: UniversalExtend[]
 }
 
 export const useTagStore = defineStore('tag', {